Complete DCT/iDCT. Add Simple Log class. Modified Watch class interface. Modified function names. And some fix.

+++ src/utils/cuda/cuda_timer.cpp
@@ -59,7 +59,7 @@ namespace util {
 			cudaEventCreate(&m_end);
 		}
 
-		int CudaStopWatch::getLapCount() const {
+		size_t CudaStopWatch::getLapCount() const {
 			return m_lap.size();
 		}
